refactor(location-bubble): seperate child components into seperate files
