--- packages/ui/src/components/Generic/LocationBubble/index.ts
@@ -1,1 +1,0 @@
-export * from "./LocationBubble"
