+++ packages/ui/src/components/Generic/LocationBubble/index.tsx
@@ -1,0 +1,125 @@
+"use client"
+import {
+  Box,
+  Center,
+  Motion,
+  useDisclosure,
+  useMotionValue,
+  useReducedMotion,
+  useTime,
+  useTransform,
+} from "@yamada-ui/react"
+import type { StaticImport } from "next/dist/shared/lib/get-img-props"
+import { useRef, useState } from "react"
+import { LocationBubbleCircle } from "./LocationBubbleCircle"
+import { LocationBubbleDesktopCard } from "./LocationBubbleDesktopCard"
+import { LocationBubbleMobileCard } from "./LocationBubbleMobileCard"
+
+/**
+ * Props for the LocationBubble component.
+ */
+export interface LocationBubbleProps {
+  locationImage: string | StaticImport
+  locationTitle: string
+  locationDetails?: string
+  locationTimes?: Record<string, string>
+  buttonLink?: string
+}
+
+/**
+ *
+ * Location Bubble component that displays a location with an floating bubble effect and provides more details on hover or click.
+ *
+ * @param locationImage - The image source for the location, can be a URL or a static import.
+ * @param locationTitle - The title of the location.
+ * @param locationDetails - Optional details about the location, such as address or description.
+ * @param locationTimes - Optional object containing days and times for the location, e.g.,
+ * `{ Tuesday: "7:30pm - 10pm", Friday: "7:30pm - 10pm" }`.
+ * @param buttonLink - Optional link for a button to learn more about the location, defaults to `"#"`.
+ * @returns A `LocationBubble` component that displays a floating circular bubble, and expands to show more details on hover or click.
+ */
+export const LocationBubble = ({
+  locationImage,
+  locationTitle,
+  locationDetails,
+  locationTimes,
+  buttonLink = "#",
+}: LocationBubbleProps) => {
+  const [hovering, setHovering] = useState(false)
+  const hoverDebounce = useRef<NodeJS.Timeout | null>(null)
+
+  const { open, onOpen, onClose } = useDisclosure()
+
+  const shouldReduceMotion = useReducedMotion()
+  const fallback = useMotionValue(0)
+  const animatedTime = useTime()
+  const time = shouldReduceMotion ? fallback : animatedTime
+  const direction = Math.random() > 0.5 ? 1 : -1
+  const xRadius = useTransform(time, (t) => Math.cos(t / 2000) * 25)
+  const yRadius = useTransform(time, (t) => Math.sin(t / 2000) * 25)
+  const bubbleXValue = useTransform(time, (t) => Math.cos(t / 3500) * xRadius.get() * direction)
+  const bubbleX = shouldReduceMotion ? fallback : bubbleXValue
+  const bubbleYValue = useTransform(time, (t) => Math.sin(t / 3500) * yRadius.get() * direction)
+  const bubbleY = shouldReduceMotion ? fallback : bubbleYValue
+
+  return (
+    <>
+      <Box data-testid="location-bubble" height={468} position="relative" width={348}>
+        {hovering ? (
+          <Motion
+            data-testid="location-bubble-desktop-card-wrapper"
+            onHoverEnd={() => {
+              if (hoverDebounce.current) {
+                clearTimeout(hoverDebounce.current)
+                hoverDebounce.current = null
+              }
+              setHovering(false)
+            }}
+          >
+            <LocationBubbleDesktopCard
+              buttonLink={buttonLink}
+              locationDetails={locationDetails}
+              locationImage={locationImage}
+              locationTimes={locationTimes}
+              locationTitle={locationTitle}
+            />
+          </Motion>
+        ) : (
+          <Center height="100%" width="100%">
+            <Motion
+              data-testid="location-bubble-circle-trigger"
+              onClick={onOpen}
+              onHoverEnd={() => {
+                if (hoverDebounce.current) {
+                  clearTimeout(hoverDebounce.current)
+                  hoverDebounce.current = null
+                }
+                setHovering(false)
+              }}
+              onHoverStart={() => {
+                hoverDebounce.current = setTimeout(() => {
+                  setHovering(true)
+                }, 100)
+              }}
+              style={{
+                x: bubbleX,
+                y: bubbleY,
+              }}
+            >
+              <LocationBubbleCircle locationImage={locationImage} locationTitle={locationTitle} />
+            </Motion>
+          </Center>
+        )}
+      </Box>
+      <LocationBubbleMobileCard
+        buttonLink={buttonLink}
+        locationDetails={locationDetails}
+        locationImage={locationImage}
+        locationTimes={locationTimes}
+        locationTitle={locationTitle}
+        onClose={onClose}
+        open={open}
+      />
+    </>
+  )
+}
