+++ packages/ui/src/components/Generic/index.ts
@@ -1,3 +1,3 @@
 export * from "./AboutUsCard/AboutUsCard"
 export * from "./Footer"
-export * from "./LocationBubble/LocationBubble"
+export * from "./LocationBubble"
