+++ packages/ui/src/test-config/mocks/LocationBubble.mock.ts
@@ -2,3 +2,9 @@ export const LOCATION_BUBBLE_TEST_CONSTANTS = {
   locationImage: "https://placehold.co/300x200/png",
   locationTitle: "Uoa Recreation Centre",
 }
+
+export const LOCATION_BUBBLE_TEST_CONSTANTS_MOBILE = {
+  ...LOCATION_BUBBLE_TEST_CONSTANTS,
+  open: true,
+  onClose: () => {},
+}
